chore: move some functions, reorder models
